feat: creates the 'infiniteScroll' method
refactor: separates the components
