+++ src/pages/Feed/feed.module.scss
@@ -4,84 +4,3 @@
   width: 100%;
   gap: 1.5rem;
 }
-
-.form {
-  max-width: calc(100% - 3rem);
-  width: 800px;
-  border-color: var(--gray-200);
-}
-
-.label {
-  margin-bottom: 1.5rem;
-}
-
-.submit {
-  padding: .5rem 2.45rem;
-}
-
-.post {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: flex-start;
-  max-width: 100%;
-}
-
-.header {
-  display: flex;
-  width: 100%;
-  align-items: center;
-  gap: 1rem;
-  padding: 1.5rem;
-  color: var(--white-100);
-  background-color: var(--blue-100);
-  border-top-right-radius: var(--border-radius-box);
-  border-top-left-radius: var(--border-radius-box);
-}
-
-.title {
-  margin-right: auto;
-}
-
-.action {
-  transition: color, transform .2s;
-
-  &:hover {
-    color: var(--white-200);
-    transform: scale(1.1);
-  }
-}
-
-.body {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  gap: 1rem;
-  padding: 1.5rem;
-  font-size: 1.12rem;
-  border: var(--border-gray-200);
-  border-top: none;
-  border-bottom-right-radius: var(--border-radius-box);
-  border-bottom-left-radius: var(--border-radius-box);
-}
-
-.username,
-.timestamp {
-  color: var(--gray-100);
-}
-
-.username {
-  font-weight: 700;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.inlineButtons {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  gap: 1rem;
-}
