+++ src/pages/FormPost/formpost.d.ts
@@ -1,8 +1,4 @@
-export interface iPostProps {
-  id: string;
-  title: string;
-  content: string;
-}
+import type { iPostProps } from '../Post/post';
 
 export interface iFormPostProps extends Partial<iPostProps> {
   onCancel?: () => void;
