+++ src/pages/FormPost/formpost.module.scss
@@ -1,0 +1,20 @@
+.form {
+  max-width: calc(100% - 3rem);
+  width: 800px;
+  border-color: var(--gray-200);
+}
+
+.label {
+  margin-bottom: 1.5rem;
+}
+
+.submit {
+  padding: .5rem 2.45rem;
+}
+
+.inlineButtons {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 1rem;
+}
