+++ src/pages/Post/post.d.ts
@@ -1,0 +1,7 @@
+export interface iPostProps {
+  id: number;
+  username: string;
+  title: string;
+  content: string;
+  created_datetime: Date;
+}
