+++ src/pages/index.ts
@@ -1,3 +1,5 @@
 export * from './Feed';
+export * from './FormPost';
 export * from './Login';
+export * from './Post';
 export * from './ProtectedRoute';
